backend/requirements.txt
acme==0.31.0
asgiref==3.3.4
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
blinker==1.4
certbot==0.31.0
certbot-apache==0.31.0
certifi==2018.1.18
cffi==1.14.5
chardet==3.0.4
click==6.7
cloud-init==21.2
colorama==0.3.7
command-not-found==0.3
ConfigArgParse==0.11.0
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.7
defusedxml==0.7.1
distro-info===0.18ubuntu0.18.04.1
Django==3.2.4
django-allauth==0.44.0
django-cors-headers==3.7.0
django-dotenv==1.4.2
django-sslserver==0.22
djangorestframework==3.12.4
ec2-hibinit-agent==1.0.0
future==0.15.2
hibagent==1.0.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
isodate==0.6.0
Jinja2==2.10
josepy==1.1.0
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
MarkupSafe==1.0
mock==2.0.0
ndg-httpsclient==0.4.4
netifaces==0.10.4
oauthlib==3.1.1
PAM==0.4.2
parsedatetime==2.4
pbr==3.1.1
pexpect==4.2.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pycrypto==2.6.1
pygobject==3.26.1
PyICU==1.9.8
PyJWT==2.1.0
pyOpenSSL==17.5.0
pyRFC3339==1.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.6
python-augeas==0.5.0
python-debian==0.1.32
python3-openid==3.2.0
pytube==10.9.2
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.16.0
sos==4.1
spotipy==2.18.0
sqlparse==0.4.1
ssh-import-id==5.7
systemd-python==234
Twisted==17.9.0
typing-extensions==3.10.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.6
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2
